% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analisisdiferencias.r
\name{analisisDiferencias}
\alias{analisisDiferencias}
\title{Compara las diferencias entre dos bases de datos}
\usage{
analisisDiferencias(x1, x2, vars.rel = colnames(x1), vars.report = x1[, 1],
  verbose = T)
}
\arguments{
\item{x1}{Base de datos uno}

\item{x2}{Base de datos dos}

\item{vars.rel}{nombre de columnas a reportar}

\item{vars.report}{vector o matriz de ids a mantener}
}
\description{
Compara las diferencias entre dos bases de datos
}
