% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_fa.r
\name{automatic_fa}
\alias{automatic_fa}
\title{Automatic Factor Analysis, to delete unnecesary factors and reageants}
\usage{
automatic_fa(x, nfactor, cut.t = 0.3, check.crossloads = T, ...)
}
\arguments{
\item{x}{data.frame}

\item{nfactor}{Initial number of factors}

\item{check.crossloads}{if the model try to eliminate crossloads. You should
disable this on biquartimin or other type of rotations}

\item{cut}{factor loading threshold}
}
\description{
La idea es seleccionar automáticamente el número correcto de factores
e ítems, a partir de una solución inicial.
Para ello, el criterio es que cada ítem pertenezca a un factor, al menor,
definido como una carga superior a un umbral.
No me opo  ndré a las cargas cruzadas, por hoy al menos
Las reglas son:
- Si un factor queda sin ítem, se selecciona un número inferior de factor
- si hay un ítem que no carga en un factor, se elimina. Se elimina primero el ítem
  con menor comunalidad
}
