% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.ml.f.r
\name{calculate.ml.f}
\alias{calculate.ml.f}
\title{Calculate likehood function for a factorial solution}
\usage{
calculate.ml.f(s.c, model = NULL, ll = NULL)
}
\arguments{
\item{s.c}{sample covariance matrix}

\item{model}{estimated model}

\item{ll}{optional loadings}
}
\description{
Calculate likehood function for a factorial solution
}
