% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.assignment.scale.r
\name{check.asignment.pair}
\alias{check.asignment.pair}
\title{Compruebo si un ítem tiene mayor carga en otro factor que en el propio
para ello, voy por pares de escalas (divide et impera)
y veo si al generar carga cruzada, su relación es mayor con el
otro ítem que con el que se supone que corresponde
Partimos con el caso simple (pares)
y de ahí creamos el general}
\usage{
check.asignment.pair(x, key)
}
\description{
Compruebo si un ítem tiene mayor carga en otro factor que en el propio
para ello, voy por pares de escalas (divide et impera)
y veo si al generar carga cruzada, su relación es mayor con el
otro ítem que con el que se supone que corresponde
Partimos con el caso simple (pares)
y de ahí creamos el general
}
