% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mplusUtils.r
\name{cleanMplusName}
\alias{cleanMplusName}
\title{Clean names to acceptable for Mplus}
\usage{
cleanMplusName(x)
}
\arguments{
\item{x}{string}

\item{a}{clean string}
}
\description{
Clean names to acceptable for Mplus
}

