% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa2xlsx.r
\name{cluster2factor}
\alias{cluster2factor}
\title{Return the assigned factor to a given solution}
\usage{
cluster2factor(x, cut)
}
\arguments{
\item{x}{fa}

\item{cut}{threshold}
}
\description{
Return the assigned factor to a given solution
}
