% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareBy.r
\name{compareBy}
\alias{compareBy}
\title{Generate a canonical description table to describe multiple variables for a
classification criterium
Includes:
- Mean and sd for each group
- Report of normality of each grupo
- Test of differences, adapted to normality of each group
- Effect size of differences: d for two groups, f for three or more}
\usage{
compareBy(
  vars,
  g,
  varnames = colnames(vars),
  use.2.bm = FALSE,
  use.3.eta2 = FALSE,
  test.type = "variable"
)
}
\arguments{
\item{vars}{a data.frame with variables}

\item{g}{vector with groups to divide the information into}

\item{test.type}{"variable" for contingent on normality test, "parametric" for
t and ANOVA and "nonparametric" for U and Kruskall-Wallis}

\item{use.2.bm:}{Use brunner.munzel.test on variables with two groups}

\item{use.3.eta2:}{Use eta squared, instead of <code>f</code>, on comparison of
three or more groups}
}
\description{
Generate a canonical description table to describe multiple variables for a
classification criterium
Includes:
- Mean and sd for each group
- Report of normality of each grupo
- Test of differences, adapted to normality of each group
- Effect size of differences: d for two groups, f for three or more
}
