% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareDatabase.r
\name{compareDatabases}
\alias{compareDatabases}
\title{Compare differences between two databases}
\usage{
compareDatabases(x1, x2, vars.rel = colnames(x1), vars.report = x1[, 1])
}
\arguments{
\item{x1}{Base de datos uno}

\item{x2}{Base de datos dos}

\item{vars.rel}{nombre de columnas a reportar}

\item{vars.report}{vector o matriz de ids a mantener}
}
\description{
Compare differences between two databases
}

