% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corAsPsych.r
\name{corAsPsych}
\alias{corAsPsych}
\title{Returns a data.frame with correlations like the psych uses}
\usage{
corAsPsych(x, delete.lower = T, round.digits = 2, ...)
}
\arguments{
\item{x}{a data.frame}

\item{delete.lower}{if should delete the lower diagonal}
}
\description{
Returns a data.frame with correlations like the psych uses
}
