% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createMplusCFA.r
\name{createMplusCFA}
\alias{createMplusCFA}
\title{Create a Mplus definition for a CFA
Requires a list of dataframes}
\usage{
createMplusCFA(x, groups, directory = ".", bfile = NULL,
  invariance_factors = F, title = "CFA Analysis", difftest = NULL)
}
\arguments{
\item{list}{of dataframes}

\item{basename}{for files}
}
\description{
Create a Mplus definition for a CFA
Requires a list of dataframes
}

