% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMplusGeneric.r
\name{createMplusGeneric}
\alias{createMplusGeneric}
\title{Create a Mplus definition for Generic Analysis}
\usage{
createMplusGeneric(x, directory = ".", bfile = NULL)
}
\arguments{
\item{x}{data.frame}

\item{directory}{directory to which save data}

\item{bfile}{base filename for scripts and data}
}
\description{
Automagicly assing factor as nominal and ordered as categorical
}
