% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createMplusLGM.r
\name{createMplusLGM}
\alias{createMplusLGM}
\title{Create a Mplus definition for a (multilevel) LGM
Requires the clustering on parameter cluster
And the different variables on dataframes of the same rank}
\usage{
createMplusLGM(x, series, pred.within, pred.between, clusters,
  model.within = c("i", "s"), model.between = c("ib", "sb"), bfile = NULL)
}
\arguments{
\item{data.frame}{}

\item{list}{of time.series}

\item{variable}{for clustering}

\item{basename}{for files}
}
\description{
Create a Mplus definition for a (multilevel) LGM
Requires the clustering on parameter cluster
And the different variables on dataframes of the same rank
}

