% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.fa.r
\name{cv.fa}
\alias{cv.fa}
\title{Cross validation of a factorial analysis.}
\usage{
cv.fa(x, ftt, folds = 5, cut = 0.3, ...)
}
\arguments{
\item{x}{dataframe}

\item{ftt}{vector with number of factor to test}

\item{folds}{number of folds for cross-validation}

\item{cut}{threshold for factorial loadings}
}
\description{
Cross validation of a factorial analysis.
}
