% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDB.r
\name{def.var.to.keys}
\alias{def.var.to.keys}
\title{Create a keys matrix for a def.var definition
like the one generateDB uses.
Useful to interface def.var files with lavaan or psych packages}
\usage{
def.var.to.keys(x, v = "var")
}
\arguments{
\item{x}{def.var definition. Can contain NA}

\item{v}{Name of var variable}
}
\description{
Create a keys matrix for a def.var definition
like the one generateDB uses.
Useful to interface def.var files with lavaan or psych packages
}
