% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descripcionGeneral.r
\name{descripcion.general}
\alias{descripcion.general}
\title{Describe a set of variables 
use ggplot}
\usage{
descripcion.general(x, vars.desc = colnames(x))
}
\description{
Describe a set of variables 
use ggplot
}
