% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportSpss.r
\name{exportSPSS}
\alias{exportSPSS}
\title{Export a data.frame to SPSS, using the metadata
of another .sav}
\usage{
exportSPSS(original, df, basename, varnames = NULL)
}
\arguments{
\item{original}{object returned by \code{read.spss}}

\item{df}{The data.frame to export}

\item{basename}{Base name for datafile and codefile}

\item{varnames}{names for columns on dataset.}
}
\description{
Export a data.frame to SPSS, using the metadata
of another .sav
}
