% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan.graficar.residuos.R
\name{fa.graficar.residuos}
\alias{fa.graficar.residuos}
\title{Grafica los residuos por ítems para un análisis factorial
La idea es de Juan Luis Castillo}
\usage{
fa.graficar.residuos(x, delete.diagonal = TRUE)
}
\arguments{
\item{modelo}{fa}
}
\description{
Grafica los residuos por ítems para un análisis factorial
La idea es de Juan Luis Castillo
}
