% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fa2xlsx.r
\name{fa2xlsx}
\alias{fa2xlsx}
\title{fa2xlsx.r
Creates a xlsx with different factorial analysis side by side
Marking the loadings to adscribe to a factor}
\usage{
fa2xlsx(x, cut = 0.3, cut.diff = 0.1, order.items = TRUE)
}
\arguments{
\item{x}{a list with factorial analysis}

\item{cut.diff}{Minimal difference between loading to assign it to a specific one}

\item{order.items}{Order items according to factor and loadings}

\item{c}{Minimal loading to assign a item to a factor}
}
\description{
fa2xlsx.r
Creates a xlsx with different factorial analysis side by side
Marking the loadings to adscribe to a factor
}

