% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa2xlsx.r
\name{fa2xlsx.wb}
\alias{fa2xlsx.wb}
\title{fa2xlsx.wb
Creates a xlsx with different factorial analysis side by side
Marking the loadings to adscribe to a factor}
\usage{
fa2xlsx.wb(
  x,
  cut = 0.3,
  cut.diff = 0.1,
  order.items = "factor",
  original.factors = NULL,
  filename = NULL
)
}
\arguments{
\item{x}{a list with factorial analysis}

\item{cut}{Minimal loading to assign a item to a factor}

\item{cut.diff}{Minimal difference between loading to assign it to a specific one}

\item{order.items}{Order items according to factor and loadings}

\item{filename}{Filename of the excel file. If NULL, create a new temporary file on each iteration}
}
\value{
an Workbook object
}
\description{
fa2xlsx.wb
Creates a xlsx with different factorial analysis side by side
Marking the loadings to adscribe to a factor
}
