% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateDB.r
\name{generateDB}
\alias{generateDB}
\title{Follow a variable definition, composed of 1 and 0
and create both a database complete set
and a means dataframe.
Like scoreItems, but less cumbersome}
\usage{
generateDB(db, def, v = "var", ignore = c(), na.rm = TRUE)
}
\arguments{
\item{db}{the database}

\item{def}{the definition matrix. Should have 'v' parameter on in}

\item{v}{name of the column names}

\item{ignore}{names of columns to ignore. By default, 'n'}
}
\description{
Follow a variable definition, composed of 1 and 0
and create both a database complete set
and a means dataframe.
Like scoreItems, but less cumbersome
}
