% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitAll.r
\name{knit.all}
\alias{knit.all}
\title{Knit all documents on a directory}
\usage{
knit.all(dir = ".", type = c("word_document", "html_document", "pdf_document"))
}
\arguments{
\item{dir}{directory}

\item{type}{type of document}
}
\description{
Knit all documents on a directory
}
