% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan.graficar.residuos.R
\name{lavaan.graficar.residuos}
\alias{lavaan.graficar.residuos}
\title{Grafica los residuos por ítem de un modelo en lavaan}
\usage{
lavaan.graficar.residuos(x)
}
\arguments{
\item{x}{lista de modelos lavaan}
}
\description{
Grafica los residuos por ítem de un modelo en lavaan
}
