% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaan2xlsx.r
\name{lavaan2xlsx}
\alias{lavaan2xlsx}
\title{Generate a xlsx file for a lavaan object}
\usage{
lavaan2xlsx(x, extra.matrix.f = NULL, order.matrix.f = NULL)
}
\arguments{
\item{x}{lavaan object}

\item{extra.matrix.f}{extra matrix to add to x object}

\item{order.matrix.f}{name of columns to which order the cor matrix}
}
\value{
a openxlsx object
}
\description{
Generate a xlsx file for a lavaan object
}
\examples{
\dontrun{
 library(psych)
 library(lavaan)
 data(Dwyer)
 cfa.1<-cfa("g=~V1+V2+V3+V4+V5+V6+V7+V8",sample.cov=Dwyer,sample.nobs=1000)
 lavaan2xlsx(cfa.1)
 }
}
