% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmtoxlsx.r
\name{lmToXlsx}
\alias{lmToXlsx}
\title{Generate a standard APA table for linear models on Excel}
\usage{
lmToXlsx(x, nested = T)
}
\arguments{
\item{nested}{if TRUE, presents R² and F for difference betweeen models}

\item{list}{of lm models}
}
\value{
an openxlsx object. You could use \code{OpenXL} to browse on spreadsheet software
}
\description{
Generate an Excel workbook with one or more lineal
models.
}
