% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.mixed.ancova.analysis.r
\name{mi.mixed.ancova.analysis.mitml}
\alias{mi.mixed.ancova.analysis.mitml}
\title{Calcular ancovas de manera general para modelos mixtos, en una base pre-establecida}
\usage{
mi.mixed.ancova.analysis.mitml(
  mi,
  v,
  mixed_model,
  subset = 1:nrow(complete(mi)),
  trans = "I",
  covariates = NULL,
  quad = F,
  extra = NULL,
  test_method = "D2",
  change.base = F
)
}
\arguments{
\item{mi}{mice multiple imputation database}

\item{v}{Vector of variables to perform ancova. First es pre-test, second is post-test and third is experimental}

\item{subset}{Subset of data to calculate within. Same as subset on lm}

\item{trans}{Transformation of @v, to use inside regression methods}

\item{covariates}{Other covariates}

\item{quad}{Use quadratic regression on pre-value}

\item{extra}{test extra variables, one on one.}

\item{test_method}{Method to use on \link{mitml::testModels} param method}

\item{change_base}{In case of moderation, change base name for ancova}
}
\description{
Calcular ancovas de manera general para modelos mixtos, en una base pre-establecida
}
