% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descripcionGeneral.r
\name{n.y.p}
\alias{n.y.p}
\title{For a specific categoric variable, presents frequency and percentage}
\usage{
n.y.p(x, ...)
}
\description{
For a specific categoric variable, presents frequency and percentage
}
