% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omg.r
\name{omg}
\alias{omg}
\title{Run an OMG Analysis,}
\usage{
omg(x, keys, partialize = NULL, ...)
}
\arguments{
\item{x}{matrix or data.frame}

\item{keys}{keys of assigment of}

\item{...}{arguments to cor.corrected}
}
\value{
a data.frame

  \describe{
  \item{item}{Item name}
  \item{assigned.name}{Name of scale (came from keys)}
  \item{cor.assigned}{Corrected correlation of item to assigned scale}
  \item{max.name}{Name of scale for maximum correlation}
  \item{cor.max}{Maximum correlation to any scale}
  \item{diff}{Difference between assigned and maximum correlation.}
  }
}
\description{
As Stuive, Kiers & Timmerman(2009) said,"The OMG method considers for each item the observed correlations between the item and each subtest, where each subtest is an unweighted sumscore of items. These correlations are corrected for two sources of spuriously high correlations, namely, self-correlation and test length[...] An assignment of an item to a subtest is considered to be correct when the corrected subtest-item correlation is highest for the subtest the item is supposed to belong to and incorrect otherwise" (p.950)
}
\examples{
library(qgraph)
data(big5)
data(big5groups)
big.keys<-sapply(big5groups,function(x) {xx<-numeric(ncol(big5));xx[x]<-1;xx})
rownames(big.keys)<-colnames(big5)
omg(big5,big.keys)

}
\seealso{
\code{\link{cor.corrected}} for corrected correlation for self-correlation and scales lengths.
}
