% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omg.r
\name{omg.xlsx}
\alias{omg.xlsx}
\title{Create a xlsx for an omg analysis}
\usage{
omg.xlsx(x, keys, order.keys = T, ...)
}
\arguments{
\item{x}{matrix or data.frame}

\item{keys}{keys of assigment of}

\item{...}{arguments to cor.corrected}
}
\value{
A Workbook object
}
\description{
Create a xlsx for an omg analysis
}
