% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplusUtils.r
\name{parlistMplus}
\alias{parlistMplus}
\title{Generate a 'at' ARE 'X1,X2'
Useful for nominal, categorical and everything related}
\usage{
parlistMplus(at, l)
}
\arguments{
\item{at}{Name of category}

\item{l}{list of variables}
}
\value{
a string
}
\description{
Generate a 'at' ARE 'X1,X2'
Useful for nominal, categorical and everything related
}
