% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotModeration.r
\name{plotModeration}
\alias{plotModeration}
\title{Plot a moderation chart.}
\usage{
plotModeration(x, y, m, k = NULL, h = NULL)
}
\arguments{
\item{x}{Predictor}

\item{y}{Variable to predict}

\item{m}{Moderator}

\item{k}{Number of cases to create a windows}

\item{h}{}
}
\value{
ggplot
}
\description{
Moderation plot allows to evaluate if a moderation is lineal, polynomial or have another form.
Two methods to calculate the moderation are used:
- Windows: Windows of size k are taken from the sample and on each the regression is calculated
- Kernel: A weigthed regression is calculated for each point, using a gaussian kernel, using the
         parameter h as bandwidth
}
\details{
You could play with k and h parameters, until you have a good approximation to data
}
