% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentaciones.R
\name{presentar.cfa}
\alias{presentar.cfa}
\title{Presenta resultados de CFA de forma estándarde forma estándar}
\usage{
presentar.cfa(x, scaled = TRUE)
}
\arguments{
\item{x}{modelo lavaan}

\item{scaled}{si los resultados están escalados}
}
\description{
Presenta resultados de CFA de forma estándarde forma estándar
}
