% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentaciones.R
\name{presentar.omega}
\alias{presentar.omega}
\title{Presenta resultados de un omega}
\usage{
presentar.omega(x)
}
\arguments{
\item{x}{resultado de omega u omegaSem}
}
\description{
Presenta resultados de un omega
}
