% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automatic_fa.r
\name{print.automatic_fa}
\alias{print.automatic_fa}
\title{Print results of automatic_fa}
\usage{
\method{print}{automatic_fa}(x)
}
\arguments{
\item{automatic_fa}{class}
}
\description{
Print results of automatic_fa
}
