% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros.r
\name{ros.from.maa}
\alias{ros.from.maa}
\title{Based on mi.mixed.ancova.analysis
generates a ros object}
\usage{
ros.from.maa(ca, moderador, se)
}
\arguments{
\item{ca}{Modelo desde un calcular ancova propiop}

\item{moderador}{la variable moderadora (usualmente el pre) TAL CUAL APARECE EN EL MODELO}

\item{se}{simple slopes -> puntos donde se define el efecto}
}
\description{
Based on mi.mixed.ancova.analysis
generates a ros object
}
