% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ros.r
\name{ros}
\alias{ros}
\title{Region of significance analysis}
\usage{
ros(model, predictor, moderador, points = NULL, simple_slopes = NULL,
  alpha = 0.05)
}
\arguments{
\item{predictor}{Name of predictor variable, as appears on coef calling}

\item{moderador}{Name of moderator variable, as appears on coef calling}

\item{points}{Matrix, with first columns values for predictor and second for moderator}

\item{simple_slopes}{Vector with moderator values}

\item{alpha}{Level of significance for test}

\item{lm}{or lmer model}
}
\description{
Region of significance analysis
}
