% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplusUtils.r
\name{writeDataFileMplus}
\alias{writeDataFileMplus}
\title{Write data file in free format for Mplus}
\usage{
writeDataFileMplus(x, filename)
}
\arguments{
\item{x}{a data.frame or matrix}

\item{filename}{Filename}
}
\description{
Write data file in free format for Mplus
}
