% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compareBy.r
\name{compareBy}
\alias{compareBy}
\title{Generate a canonical description table to describe multiple variables for a
classification criterium
Includes:
- Mean and sd for each group
- Report of normality of each grupo
- Test of differences, adapted to normality of each group
- Effect size of differences: d for two groups, f for three or more}
\usage{
compareBy(vars, g, varnames = colnames(vars), use.2.bm = FALSE)
}
\arguments{
\item{vars}{a data.frame with variables}

\item{g}{groups to divide the information into}

\item{use.2.bm:}{Use brunner.munzel.test on variables with two groups}
}
\description{
Generate a canonical description table to describe multiple variables for a
classification criterium
Includes:
- Mean and sd for each group
- Report of normality of each grupo
- Test of differences, adapted to normality of each group
- Effect size of differences: d for two groups, f for three or more
}

