% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareDichotomicBy.r
\name{compareDichotomicBy}
\alias{compareDichotomicBy}
\title{Generate a canonical description table to describe multiple dichotomic variables
By default, the first level is the 'correct' answer.
Includes: 
- Proportion of correct 
- Chi-Square test
- Effect size: Cramer V}
\usage{
compareDichotomicBy(vars, g, varnames = colnames(vars), success = NULL)
}
\arguments{
\item{vars}{a data.frame with variables}

\item{g}{groups to divide the information into}
}
\description{
Generate a canonical description table to describe multiple dichotomic variables
By default, the first level is the 'correct' answer.
Includes: 
- Proportion of correct 
- Chi-Square test
- Effect size: Cramer V
}
