% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMplusCFA.r
\name{createMplusCFA}
\alias{createMplusCFA}
\title{Create a Mplus definition for a CFA
Requires a list of dataframes}
\usage{
createMplusCFA(x, groups, directory = ".", bfile = NULL,
  invariance_factors = F, title = "CFA Analysis", difftest = NULL)
}
\arguments{
\item{x}{list of dataframes}

\item{groups}{vector of assignation to groups}

\item{directory}{directory where to store results}

\item{bfile}{base name for model files}

\item{title}{Title of analysis}

\item{difftest}{Calculate difftest with another model}

\item{invariance_factor}{Calculate invariance between groups}
}
\description{
Create a Mplus definition for a CFA
Requires a list of dataframes
}
