% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createMplusCFA}
\alias{createMplusCFA}
\title{Create a Mplus definition for a CFA
Requires a list of dataframes}
\usage{
createMplusCFA(x, bfile)
}
\arguments{
\item{list}{of dataframes}

\item{basename}{for files}
}
\description{
Create a Mplus definition for a CFA
Requires a list of dataframes
}

