% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMplusLCA.r
\name{createMplusLCA}
\alias{createMplusLCA}
\title{Create a Mplus definition for Latent Class Analysis}
\usage{
createMplusLCA(x, classes, lc.vars = colnames(x), predictors,
  bfile = NULL)
}
\arguments{
\item{x}{data.frame with data}

\item{classes}{number of latent classes}

\item{lc.vars}{list of variables to perform latent class. By default: all}

\item{predictors}{variable for regression on latent class. By default: none}

\item{bfile}{basename for files}
}
\description{
Create a Mplus definition for Latent Class Analysis
}
