% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createMplusLCA.r
\name{createMplusLCA}
\alias{createMplusLCA}
\title{Create a Mplus definition for Latent Class Analysis}
\usage{
createMplusLCA(x, classes, lc.vars = colnames(x), predictors, bfile = NULL)
}
\arguments{
\item{data.frame}{}

\item{list}{of variables to perform latent class. By default: all}

\item{variable}{for regression on latent class. By default: none}

\item{basename}{for files}
}
\description{
Create a Mplus definition for Latent Class Analysis
}

