% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMplusLGM.r
\name{createMplusLGM}
\alias{createMplusLGM}
\title{Create a Mplus definition for a (multilevel) LGM
Requires the clustering on parameter cluster
And the different variables on dataframes of the same rank}
\usage{
createMplusLGM(x, series, pred.within, pred.between, clusters,
  model.within = c("i", "s"), model.between = c("ib", "sb"),
  bfile = NULL)
}
\arguments{
\item{x}{data.frame}

\item{series}{list of time.series}

\item{pred.within}{Predictors for within level}

\item{pred.between}{Predictors for between levels}

\item{clusters}{name of variable for clustering}

\item{model.within}{by default c("i","s")}

\item{model.between}{by default c("ib","sb")}

\item{bfile}{basename for files}
}
\description{
Create a Mplus definition for a (multilevel) LGM
Requires the clustering on parameter cluster
And the different variables on dataframes of the same rank
}
