% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fa2xlsx.r
\name{fa2xlsx}
\alias{fa2xlsx}
\title{fa2xlsx
View Creates a xlsx with different factorial analysis side by side
Marking the loadings to adscribe to a factor.}
\usage{
fa2xlsx(..., cut = 0.3, cut.diff = 0.1, order.items = "factor",
  original.factors = NULL, filename = NULL)
}
\arguments{
\item{...}{one or more factorial analysis}

\item{cut}{Minimal loading to assign a item to a factor}

\item{cut.diff}{Minimal difference between loading to assign it to a specific one}

\item{order.items}{A string, according to specification. 
"factor" orders items according to new factors and loads 
"o.factor" orders according original factors and loads
"o.factor-items" orders according to factors and items names}

\item{original.factors}{Vector of names for original factors. Useful for target rotations}

\item{filename}{Filename of the excel file. If NULL, create a new temporary file on each iteration}
}
\description{
fa2xlsx
View Creates a xlsx with different factorial analysis side by side
Marking the loadings to adscribe to a factor.
}
