% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.descriptives.r
\name{mi.descriptives}
\alias{mi.descriptives}
\title{Calculate mean, standard deviation, means' confidence intervales for missing data}
\usage{
mi.descriptives(
  mi,
  subset = 1:nrow(mice::complete(mi)),
  FUN = I,
  variables = 1:ncol(FUN(mice::complete(mi)))
)
}
\description{
Calculate mean, standard deviation, means' confidence intervales for missing data
}
