% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presentaciones.R
\name{presentar.kmo.bartlett}
\alias{presentar.kmo.bartlett}
\title{Presenta resultados de KMO / Bartlett de forma estándar}
\usage{
presentar.kmo.bartlett(x, base)
}
\arguments{
\item{x}{matriz de correlaciones}

\item{base}{base de datos en bruto}
}
\description{
Presenta resultados de KMO / Bartlett de forma estándar
}
