% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplusUtils.r
\name{typeConvMplus}
\alias{typeConvMplus}
\title{variable list
you could set as free (*), with a fixed value (<X>) or related to a variable ( (label) )
they could be part of the same definition, or separated (using ;)}
\usage{
typeConvMplus(t)
}
\description{
variable list
you could set as free (*), with a fixed value (<X>) or related to a variable ( (label) )
they could be part of the same definition, or separated (using ;)
}
